Separate construction and encoding of export
